pyproject.toml
[project]
name = "querysight"
version = "0.1.0"
description = "QuerySight: ClickHouse Log-Driven dbt Project Enhancer"
requires-python = ">=3.11"
dependencies = [
    "clickhouse-driver>=0.2.9",
    "openai>=1.57.4",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "reportlab>=4.2.5",
    "streamlit>=1.41.1",
    "trafilatura>=2.0.0",
    "twilio>=9.4.1",
]
